Move runtime user information function to io module.
