+++ src/msmhelper/io.py
@@ -2,15 +2,20 @@
 """Input and output text files.
 
 BSD 3-Clause License
-Copyright (c) 2019-2020, Daniel Nagel
+Copyright (c) 2019-2023, Daniel Nagel
 All rights reserved.
 
 """
+import datetime
+import getpass  # noqa: SC100 # get user name with getpass.getuser()
+import os
+import platform  # get pc name with platform.node()
 import sys
 
 import numpy as np
 import pandas as pd
 
+import __main__ as main
 from msmhelper import tools
 
 
@@ -107,7 +112,7 @@ def savetxt(file_name, array, header=None, fmt='%.5f'):  # noqa: WPS323
 
     """
     # prepare header comments
-    RUI = tools.get_runtime_user_information()
+    RUI = _get_runtime_user_information()
 
     header_comment = (
         'This file was generated by {script_dir}/{script_name}:\n{args}' +
@@ -221,3 +226,39 @@ def open_limits(data_length, limits_file=None):
         raise ValueError('Limits are inconsistent to data.')
 
     return limits
+
+
+def _get_runtime_user_information():
+    r"""Get user runtime information.
+
+    !!! warning
+        For python 3.5 or lower the date is not formatted and contains
+        microscends.
+
+    Returns
+    -------
+    RUI : dict
+        Holding username in 'user', pc name in 'pc', date of execution 'date',
+        path of execution 'script_dir' and name of execution main file
+        'script_name'. In case of interactive usage, script_name is 'console'.
+
+    """
+    try:
+        script_dir, script_name = os.path.split(
+            os.path.abspath(main.__file__),  # noqa: WPS609
+        )
+    except AttributeError:  # pragma: no cover
+        script_dir, script_name = '', 'console'
+
+    # get time without microseconds
+    date = datetime.datetime.now()
+    if sys.version_info >= (3, 6):
+        date = date.isoformat(sep=' ', timespec='seconds')
+
+    return {
+        'user': getpass.getuser(),
+        'pc': platform.node(),
+        'date': date,
+        'script_dir': script_dir,
+        'script_name': script_name,
+    }
